Rename /base -> /codec
