+++ packages/api-codec/src/AccountId.ts
@@ -9,8 +9,8 @@ import isHex from '@polkadot/util/is/hex';
 import isU8a from '@polkadot/util/is/u8a';
 import u8aToU8a from '@polkadot/util/u8a/toU8a';
 
-import CodecU8a from './base/U8a';
-import CodecU8aFixed from './base/U8aFixed';
+import CodecU8a from './codec/U8a';
+import CodecU8aFixed from './codec/U8aFixed';
 
 // A wrapper around an AccountId/PublicKey representation. Since we are dealing with
 // underlying PublicKeys (32 bytes in length), we extend from CodecU8aFixed which is
