+++ packages/api-codec/src/H256.ts
@@ -2,8 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import CodecU8a from './base/U8a';
-import CodecU8aFixed from './base/U8aFixed';
+import CodecU8a from './codec/U8a';
+import CodecU8aFixed from './codec/U8aFixed';
 
 // Hash containing 256 bits (32 bytes), typically used in blocks, extrinsics and
 // as a sane default for fixed-length hash representations.
