+++ packages/api-codec/src/H512.ts
@@ -2,8 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import CodecU8a from './base/U8a';
-import CodecU8aFixed from './base/U8aFixed';
+import CodecU8a from './codec/U8a';
+import CodecU8aFixed from './codec/U8aFixed';
 
 // Hash containing 512 bits (64 bytes), typically used for signatures
 export default class H512 extends CodecU8aFixed {
