+++ packages/api-codec/src/Metadata.decoded.json
@@ -676,7 +676,7 @@
             ]
           },
           {
-            "name": "TransactionCodecBaseFee",
+            "name": "TransactionBaseFee",
             "modifier": 2,
             "type": "T::Balance",
             "documentation": [
