+++ packages/api-codec/src/String.ts
@@ -6,8 +6,8 @@ import u8aFromUtf8 from '@polkadot/util/u8a/fromUtf8';
 import u8aToUtf8 from '@polkadot/util/u8a/toUtf8';
 import u8aConcat from '@polkadot/util/u8a/concat';
 
-import CodecBase from './base/Base';
-import Length from './base/Length';
+import CodecBase from './codec/Base';
+import Length from './codec/Length';
 
 // This is a string wrapper, along with the length. It is used both for strings as well
 // as stuff like documentation.
