+++ packages/api-codec/src/codec/U8a.ts
@@ -5,13 +5,13 @@
 import u8aToHex from '@polkadot/util/u8a/toHex';
 import toU8a from '@polkadot/util/u8a/toU8a';
 
-import Base from './Base';
+import CodecBase from './Base';
 
 // A CodecU8a. A basic wrapper around Uint8Array, with no frills and no fuss. It
 // wraps a Uint8Array. It does differ from other implementations wher it will
 // consume the full u8a as passed to it in fromU8a. As such it is meant to be
 // subclassed where the wrapper takes care of the actual lengths.
-export default class CodecU8a extends Base<Uint8Array> {
+export default class CodecU8a extends CodecBase<Uint8Array> {
   constructor (value: CodecU8a | string | Uint8Array = new Uint8Array()) {
     super(
       value instanceof CodecU8a
