+++ packages/api-codec/src/codec/U8aFixed.ts
